:label: SEA-19: update types to include contract and oracle
